+++ src/types/queries/epochs.ts
@@ -90,6 +90,7 @@ export interface EpochParameters {
   min_pool_cost: string;
   nonce: string;
   cost_models: Record<string, unknown> | null;
+  cost_models_raw: Record<string, unknown> | null;
   price_mem: number;
   price_step: number;
   max_tx_ex_mem: string;
@@ -123,6 +124,7 @@ export interface EpochParameters {
   drep_activity: string | null;
   registered_tx_id: number | null;
   pvtpp_security_group: number | null;
+  pvt_p_p_security_group: number | null;
   min_fee_ref_script_cost_per_byte: number | null;
 }