+++ src/types/queries/governance.ts
@@ -214,6 +214,7 @@ export interface ProposalsProposalParameters {
     drep_activity: string | null;
     registered_tx_id: number | null;
     pvtpp_security_group: number | null;
+    pvt_p_p_security_group: number | null;
     min_fee_ref_script_cost_per_byte: number | null;
   };
 }