+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `/txs/{hash}/utxos`
+  - `consumed_by_tx` field
+- `/epochs/{number}/parameters` and `/epochs/latest/parameters`
+  - `cost_models_raw` field, list variant of cost_models without name mapping
+
+### Fixed
+
+- Naming of `pvtpp_security_group` -> `pvt_p_p_security_group`, the old field is preserved but marked as deprecated.
+
 ## [2.1.1] - 2024-08-23
 
 ### Changed